Cleaned up whitespace to (more closely) match the Omni coding standards, since this old style was burning my eyes. Reordered some methods for similar reasons.
Turned a bunch of properties into plain ivars to hide the information.

Fixed a leak and a release-before-retain issue.

Removed a bunch of NSLog statements.

Deleted a bunch of template code that we don't need.

(Yeah, this should have been a bunch of tiny commits. I'm a bad, bad git user.)